math-mentor/requirements.txt
streamlit==1.31.0
google-generativeai==0.8.3
openai==1.12.0
langchain==0.1.9
langchain-community==0.0.24
chromadb==0.4.22
Pillow==10.2.0
easyocr==1.7.0
openai-whisper==20231117
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
python-dotenv==1.0.1
pydantic==2.6.1
numpy==1.26.4
pandas==2.2.0
faiss-cpu==1.7.4
opencv-python-headless==4.8.1.78
soundfile==0.12.1
